chore: move translations file and this documentation to `i18n` folder
